+++ src/main/java/org/ademun/timetableapi/repository/DayRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DayRepository extends JpaRepository<Day, Long> {
+public interface DayRepository extends JpaRepository<Day, Integer> {
 
 }
