+++ src/main/java/org/ademun/timetableapi/repository/DisciplineRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DisciplineRepository extends JpaRepository<Discipline, Long> {
+public interface DisciplineRepository extends JpaRepository<Discipline, Integer> {
 }
